Extract pagination component
