+++ src/components/index.js
@@ -8,3 +8,4 @@ export { default as Tag } from './Tag';
 export { default as Day } from './Day';
 export * from './typography';
 export Snippet from './Snippet';
+export Pagination from './Pagination';
